Potential fix for code scanning alert no. 4: Multiplication result converted to larger type

+++ bmp.c
@@ -258,7 +258,7 @@ Image* bmpLoad( strbyte* fullname, ImageOptions* image_ops, bool verbose )
 
 		if ( hdr.biCompression == BI_RLE4 ) {
 			int d, e;
-			memset( ( byte* ) image->data, 0, image->width * image->height );
+			memset( ( byte* ) image->data, 0, (size_t)image->width * (size_t)image->height );
 			for ( x = y = 0;; ) {
 				int i, f;
 				if ( ( d = zgetc( zf ) ) == EOF ) {
